Optimzie: 	1. Reorganize by separate the tests to /tests folder. 	2. Add the new API to get count of union instead of access inner field directly.
